Replace 'int' with 'size_t'

+++ map/map_tf_generator/include/map_tf_generator/uniform_random.hpp
@@ -18,12 +18,12 @@
 #include <random>
 #include <vector>
 
-std::vector<int> UniformRandom(const int max_excluded, const int n)
+std::vector<size_t> UniformRandom(const int max_excluded, const int n)
 {
   std::default_random_engine generator;
-  std::uniform_int_distribution<int> distribution(0, max_excluded - 1);
+  std::uniform_int_distribution<size_t> distribution(0, max_excluded - 1);
 
-  std::vector<int> v(n);
+  std::vector<size_t> v(n);
   for (int i = 0; i < n; ++i) {
     v[i] = distribution(generator);
   }
